+++ packages/terra-collapsible-menu-view/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+6.15.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
 6.14.0 - (October 3, 2019)
 ------------------
 ### Changed