+++ packages/terra-date-input/CHANGELOG.md
@@ -4,6 +4,16 @@ ChangeLog
 Unreleased
 ----------
 
+1.3.0 - (October 16, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+### Added
+* Added `isIncomplete` prop and stylings
+* Added `required` prop
+* Added border css styles when `isInvalid` prop is provided
+
 1.2.0 - (October 3, 2019)
 ------------------
 ### Changed